refactor: move barrier
